Rewrote and improved functions_and_methods.py
